requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
fastapi
uvicorn[standard]
passlib[argon2]
python-jose[cryptography]
python-multipart
aiofiles
